+++ src/components/NewButtons/NewButtonLabelWithScreenReaderTitle.tsx
@@ -2,8 +2,8 @@ import { FC } from "react";
 import styled from "styled-components";
 
 import typography from "../../styles/utils/typography";
-import { SpanProps } from "../SharedComponents/Typography/Span";
 
+import { SpanProps } from "@/components/SharedComponents/Typography/Span";
 import { Span } from "@/components/SharedComponents/Typography";
 import ScreenReaderOnly from "@/components/SharedComponents/ScreenReaderOnly";
