+++ src/components/OutlineHeading/OutlineHeading.tsx
@@ -4,10 +4,11 @@ import color from "../../styles/utils/color";
 import responsive, { ResponsiveValues } from "../../styles/utils/responsive";
 import { margin } from "../../styles/utils/spacing";
 import typography from "../../styles/utils/typography";
+
 import {
   HeadingProps,
   HeadingTagComponent,
-} from "../SharedComponents/Typography/Heading";
+} from "@/components/SharedComponents/Typography/Heading";
 
 export const outlineShadow = `-1px -1px 0 #000, 1px -1px 0 #000, -1px 1px 0 #000, 1px 1px 0 #000}`;
 const outlineShadowLight = `-1px -1px 0 #878787, 1px -1px 0 #878787, -1px 1px 0 #878787, 1px 1px 0 #878787}`;
