+++ src/components/SharedComponents/Button/IconButtonInner.tsx
@@ -3,7 +3,6 @@ import { useTheme } from "styled-components";
 
 import { OakColorName } from "../../../styles/theme";
 
-
 import {
   ButtonBackground,
   ButtonSize,
