+++ src/components/SharedComponents/Typography/Typography.tsx
@@ -1,7 +1,8 @@
 import styled from "styled-components";
 
 import typography, { TypographyProps } from "../../../styles/utils/typography";
-import Box, { BoxProps } from "../Box";
+
+import Box, { BoxProps } from "@/components/SharedComponents/Box";
 
 type TypographyComponent = BoxProps & TypographyProps;
 /**
