+++ src/components/pages/SubjectListing.page.tsx
@@ -2,9 +2,9 @@ import { FC } from "react";
 
 import { SubjectListingPageProps } from "../../pages/teachers/key-stages/[keyStageSlug]/subjects";
 import SubjectCardDouble from "../SubjectCardDouble/SubjectCardDouble";
-import { GridList } from "../SharedComponents/Typography/UL";
-import { GridAreaListItem } from "../SharedComponents/Typography/LI";
 
+import { GridList } from "@/components/SharedComponents/Typography/UL";
+import { GridAreaListItem } from "@/components/SharedComponents/Typography/LI";
 import { Heading } from "@/components/SharedComponents/Typography";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
 import Flex from "@/components/SharedComponents/Flex";
